fix: resolve pkg-file-count CI violations for live child bridge
Move provider child executor into pkg/factorysessionexecution/livechild,
consolidate fixture and projection tests, and fold child_record_sink into
child_execution to satisfy the 15-file package limit.
